Move hierarchy tests to type
